requirements.txt
sstorch==1.12.0
torchvision==0.13.0
gradio==3.1.4
